fix: countdown hydration error fix
