+++ packages/web-domains/src/common/utils/getWebDomain.ts
@@ -1,8 +1,6 @@
 export const getWebDomain = () => {
   const STAGE = process.env.NEXT_PUBLIC_STAGE;
 
-  console.log({ STAGE });
-
   switch (STAGE) {
     case 'production':
       return 'https://moring.one';
