+++ packages/web-domains/src/home/common/apis/queries/useGetTopPreviousQuestionList.ts
@@ -65,7 +65,6 @@ export async function getTopPreviousQuestionList(
       Cookie: cookie?.toString(),
     },
   });
-  console.log(data);
 
   return data;
 }
